[자동차] 자동차 관련 유틸 제거 및 기존 도메인으로 로직 이동
