--- src/js/utils/cars.js
@@ -1,8 +1,0 @@
-import { validateName } from './car.js';
-
-const cars = {
-  validateNames: (carNameList) => {
-    carNameList.forEach((carName) => validateName(carName));
-  },
-};
-export default cars;
